link-tests: put macho search strategy tests into one test case
