--- test/link/macho/search_paths_first/a.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
-
-char world[] = "world";
-
-char* hello() {
-  return "Hello";
-}
