--- test/link/macho/search_paths_first/main.c
@@ -1,9 +1,0 @@
-#include <stdio.h>
-
-char* hello();
-extern char world[];
-
-int main() {
-  printf("%s %s", hello(), world);
-  return 0;
-}
