+++ test/link/macho/search_strategy/a.c
