+++ test/link/macho/search_strategy/main.c
